requirements.txt
absl-py==1.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.7
astunparse==1.6.3
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.6.0
bleach==5.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.1.0
cmake==3.24.0
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dm-tree==0.1.7
entrypoints==0.4
executing==0.9.1
fastjsonschema==2.16.1
flake8==5.0.4
flatbuffers==1.12
fonttools==4.34.4
gast==0.4.0
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.47.0
h5py==3.7.0
hdbscan==0.8.28
idna==3.3
imageio==2.21.1
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.9
jsonschema==4.9.1
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-packaging==0.12.2
jupyter-scatter==0.7.1
jupyter-server==1.18.1
jupyterlab==3.4.4
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
keras==2.9.0
Keras-Preprocessing==1.1.2
keras-tuner==1.1.3
kiwisolver==1.4.4
kt-legacy==1.0.4
libclang==14.0.6
llvmlite==0.39.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.7.0
mistune==0.8.4
MulticoreTSNE==0.1
mypy==0.971
mypy-extensions==0.4.3
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.1
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
notebook==6.4.12
notebook-shim==0.1.0
numba==0.56.0
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.0
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pydot==1.4.2
pyflakes==2.5.0
Pygments==2.12.0
pynndescent==0.5.7
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-probability==0.17.0
termcolor==1.1.0
terminado==0.15.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2022.8.8
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.11.3
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
traittypes==0.2.1
typing_extensions==4.3.0
umap-learn==0.5.3
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.2.2
wget==3.2
widgetsnbextension==3.6.1
wrapt==1.14.1
zipp==3.8.1
